requirements.txt
pandas
dash
boto3
requests
python-dotenv
yfinance